TetraMesh: make TetraMesh a SpatialEntity
